requirements.txt
pandas
fastapi
streamlit
pillow
numpy
transformer